restructure of test groups
